+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -15,11 +15,11 @@ return new class extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->id();
-            $table->string('name');
-            $table->string('email')->unique();
-            $table->timestamp('email_verified_at')->nullable();
-            $table->string('password');
-            $table->rememberToken();
+            $table->string('external_id')->unique();
+            $table->string('name')->nullable();
+            $table->string('email')->nullable();
+            $table->string('password')->nullable();
+            $table->tinyInteger('status')->default(\App\Enums\UserStatusEnums::ACTIVE->value);
             $table->timestamps();
         });
     }