+++ ChangeLog.md
@@ -6,5 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
-## [0.0.1] - 2022-05-07
+## [0.0.1] - 2022-05-08
 ### Added
+- User Status Enum class
+
+### Changed
+- User Model to have external ID and Status